requirements.txt
bcrypt==3.1.7
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
cryptography==3.3.2
Django==1.7.11
enum34==1.1.10
Genshi==0.7.7
idna==2.10
ipaddress==1.0.23
paramiko==2.12.0
pip==9.0.3
protobuf==3.18.0
pycparser==2.21
pycrypto==2.6.1
PyMySQL==0.8.0
PyNaCl==1.4.0
requests==2.27.1
setuptools==39.0.1
six==1.16.0
urllib3==1.26.12
bots==3.2.0
CherryPy==7.1.0
